Set progress counts when resuming a sync.
